+++ app/src/main/java/week/on/a/plate/screens/searchRecipes/logic/SearchViewModel.kt
@@ -277,7 +277,7 @@ class SearchViewModel @Inject constructor(
                                 0,
                                 it.description.value,
                                 it.image.value,
-                                it.timer.longValue, it.start.toLong(), it.duration.toLong(), it.pinnedIngredientsInd.value
+                                it.timer.longValue, it.start, it.duration, it.pinnedIngredientsInd.value
                             )
                         },
                         link = recipe.source.value, false, LocalDateTime.now()
